Add baremetal _nostd variant
Note that product_available and vendor_available are dropped from the
existing target because they are irrelevant for a _host target (as per
the change in aosp/2763825).

Add a visibility restriction to the new variant to preserve the
limitation that this crate should not be used in general Android code.

Bug: 303192701
Test: build both variants
Change-Id: I158dab8eda7562ba634c2a9f72dfbe490609ae50